% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rmcc_cloud}
\alias{rmcc_cloud}
\title{Airborne LiDAR point cloud}
\format{
A data frame with 28668 rows and 3 variables:
\describe{
\item{X}{x coordinates}
\item{Y}{y coordinates}
\item{Z}{z coordinates}
}
}
\usage{
rmcc_cloud
}
\description{
A dataset containing a small point cloud acquired with airborne LiDAR.
}
\keyword{datasets}
